+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/api/methods/updatingmessages/EditMessageReplyMarkup.java
@@ -11,16 +11,18 @@ import org.telegram.telegrambots.exceptions.TelegramApiRequestException;
 import org.telegram.telegrambots.exceptions.TelegramApiValidationException;
 
 import java.io.IOException;
+import java.io.Serializable;
 import java.util.Objects;
 
 /**
  * @author Ruben Bermudez
  * @version 1.0
  * @brief Use this method to edit only the reply markup of messages sent by the bot or via the bot
- * (for inline bots). On success, the edited Message is returned.
+ * (for inline bots). On success, if edited message is sent by the bot, the edited Message is returned,
+ * therwise True is returned.
  * @date 10 of April of 2016
  */
-public class EditMessageReplyMarkup extends BotApiMethod<Message> {
+public class EditMessageReplyMarkup extends BotApiMethod<Serializable> {
     public static final String PATH = "editmessagereplymarkup";
 
     private static final String CHATID_FIELD = "chat_id";
@@ -99,7 +101,7 @@ public class EditMessageReplyMarkup extends BotApiMethod<Message> {
     }
 
     @Override
-    public Message deserializeResponse(String answer) throws TelegramApiRequestException {
+    public Serializable deserializeResponse(String answer) throws TelegramApiRequestException {
         try {
             ApiResponse<Message> result = OBJECT_MAPPER.readValue(answer,
                     new TypeReference<ApiResponse<Message>>(){});
@@ -109,7 +111,18 @@ public class EditMessageReplyMarkup extends BotApiMethod<Message> {
                 throw new TelegramApiRequestException("Error editing message reply markup", result);
             }
         } catch (IOException e) {
-            throw new TelegramApiRequestException("Unable to deserialize response", e);
+            try {
+                ApiResponse<Boolean> result = OBJECT_MAPPER.readValue(answer,
+                        new TypeReference<ApiResponse<Boolean>>() {
+                        });
+                if (result.getOk()) {
+                    return result.getResult();
+                } else {
+                    throw new TelegramApiRequestException("Error editing message reply markup", result);
+                }
+            } catch (IOException e2) {
+                throw new TelegramApiRequestException("Unable to deserialize response", e);
+            }
         }
     }